[ADD] #45 - Readability fixes
