+++ sources/rendering/make_minimap.c
@@ -6,15 +6,13 @@
 /*   By: sikunne <sikunne@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/01 12:42:58 by sikunne           #+#    #+#             */
-/*   Updated: 2025/07/22 22:10:46 by sikunne          ###   ########.fr       */
+/*   Updated: 2025/07/22 22:43:49 by sikunne          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../headers/cub3d.h"
-# define MMAP_SIZE 150
-# define PLAYER_SIZE 5
-# define VIEW_DISTANCE 10
 
+/** */
 static void st_hor_line(t_data *data, int x, int y, int color)
 {
 	while (++x < MMAP_SIZE)
@@ -75,9 +73,9 @@ x + steps * (data->dir_x - data->plane_x) > (double)(MMAP_SIZE - 1) || \
 y + steps * (data->dir_y - data->plane_y) < 0 || \
 y + steps * (data->dir_y - data->plane_y) > (double)(MMAP_SIZE - 1))
 			break ;
-		my_pixel_put(data->image, (int)(x + steps * (data->dir_x + data->plane_x)), (int)(y + steps * (data->dir_y + data->plane_y)), to_rgb(0, 0, 0));
-		my_pixel_put(data->image, (int)(x + steps * (data->dir_x - data->plane_x)), (int)(y + steps * (data->dir_y - data->plane_y)), to_rgb(0, 0, 0));
-		steps += 0.5;
+		my_pixel_put(data->image, (int)(x + steps * (data->dir_x + data->plane_x)), (int)(y + steps * (data->dir_y + data->plane_y)), to_rgb(0, 100, 0));
+		my_pixel_put(data->image, (int)(x + steps * (data->dir_x - data->plane_x)), (int)(y + steps * (data->dir_y - data->plane_y)), to_rgb(0, 100, 0));
+		steps += 0.25;
 	}
 }
 
@@ -86,11 +84,11 @@ static void	st_draw_player(t_data *data)
 	int		x;
 	int		y;
 
-	y = (MMAP_SIZE / 2) - (PLAYER_SIZE / 2) - 1;
-	while (++y < (MMAP_SIZE / 2) + (PLAYER_SIZE / 2))
+	y = (MMAP_SIZE / 2) - (MMAP_PLAYER_SIZE / 2) - 1;
+	while (++y < (MMAP_SIZE / 2) + (MMAP_PLAYER_SIZE / 2))
 	{
-		x = (MMAP_SIZE / 2) - (PLAYER_SIZE / 2) - 1;
-		while (++x < (MMAP_SIZE / 2) + (PLAYER_SIZE / 2))
+		x = (MMAP_SIZE / 2) - (MMAP_PLAYER_SIZE / 2) - 1;
+		while (++x < (MMAP_SIZE / 2) + (MMAP_PLAYER_SIZE / 2))
 		{
 			my_pixel_put(data->image, x, y, to_rgb(102, 102, 102));
 		}
@@ -98,62 +96,9 @@ static void	st_draw_player(t_data *data)
 	st_draw_fov(data);
 }
 
-static double	st_convert(int current, int max)
-{
-	double adjust;
-	double factor;
-
-	adjust = (double)current - (double)(max / 2);
-	factor = adjust / (double)max;
-	return (factor * VIEW_DISTANCE);
-}
-
-static int st_col(t_data *data, double posx, double posy)
-{
-	int x;
-	int y;
-
-	x = floor(posx);
-	y = floor(posy);
-	if (x < 0 || y < 0 || x > data->minfo->width - 1 || y > data->minfo->height - 1)
-		return (0);
-	if (data->minfo->grid[y][x] == '1')
-		return(to_rgb(171, 144, 101));
-	if (data->minfo->grid[y][x] == ' ')
-		return(0);
-	return(to_rgb(221, 194, 151));
-}
-
-// 14	-1
-// 32	0
-// 70	1
-
-static void st_draw_blocks(t_data *data)
-{
-	int	x;
-	int y;
-	double posx;
-	double	posy;
-	int	color;
-
-	y = -1;
-	while (++y < MMAP_SIZE)
-	{
-		x = -1;
-		while (++x < MMAP_SIZE)
-		{
-			posx = data->pos_x + st_convert(x, MMAP_SIZE);
-			posy = data->pos_y + st_convert(y, MMAP_SIZE);
-			color = st_col(data, posx, posy);
-			if (color != 0)
-				my_pixel_put(data->image, x, y, color);	
-		}
-	}
-}
-
 void	make_minimap(t_data *data)
 {
-	st_draw_blocks(data);
+	minimap_background(data);
 	st_draw_player(data);
 	st_draw_frame(data);
 }
